@@ -4,28 +4,29 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
+  equatable: ^2.0.0
   flutter:
     sdk: flutter
   flutter_bloc:
     path: ../../packages/flutter_bloc
-  equatable: ^2.0.0
 
 dev_dependencies:
+  bloc_test: ^9.0.0
   flutter_test:
     sdk: flutter
-  bloc_test: ^9.0.0
-  mocktail: ^0.2.0
+  mocktail: ^0.3.0
+  very_good_analysis: ^3.0.0
 
 dependency_overrides:
   bloc:
     path: ../../packages/bloc
-  flutter_bloc:
-    path: ../../packages/flutter_bloc
   bloc_test:
     path: ../../packages/bloc_test
+  flutter_bloc:
+    path: ../../packages/flutter_bloc
 
 flutter:
   uses-material-design: true
