@@ -4,7 +4,7 @@ import 'package:flutter_complex_list/complex_list/complex_list.dart';
 import 'package:flutter_complex_list/repository.dart';
 
 class App extends MaterialApp {
-  App({required Repository repository})
+  App({super.key, required Repository repository})
       : super(
           home: RepositoryProvider.value(
             value: repository,
